pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
(
  /(
    | venv
    | _build
    | buck-out
    | build
    | migrations
    | config
  )/
)
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 120

[tool.poetry]
name = "mspp"
version = "0.1.0"
description = ""
authors = ["Ya.Practicum Students"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.5"
dj-database-url = "^1.2.0"
python-decouple = "^3.7"
python-telegram-bot = "^20.0"
uvicorn = "^0.20.0"
python-environ = "^0.4.54"
django-mptt = "^0.14.0"
aiogoogle = "^5.1.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.0.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
black = "^22.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
